1단계 가입 컴포넌트에서 2단계 가입 컴포넌트로 포지션 선택 영역 변경
